add manage listings rte, view, funct

+++ app/Http/Controllers/ListingController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Auth;
 
 class ListingController extends Controller
 {
@@ -48,6 +49,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('success', 'Listing created successfully!');
@@ -62,6 +65,11 @@ class ListingController extends Controller
     //Update Listing
     public function update(Request $request, Listing $listing)
     {
+        //check logged in user
+        if ($listing->user_id != Auth::user()->id) {
+            abort(403, 'Unauthorized Access');
+        }
+
         //form validation
         $formFields = $request->validate([
             'title' => 'required',
@@ -82,9 +90,21 @@ class ListingController extends Controller
         return back()->with('success', 'Listing updated successfully!');
     }
 
+    //Delete Listing
     public function destroy(Listing $listing)
     {
+        //check logged in user
+        if ($listing->user_id != Auth::user()->id) {
+            abort(403, 'Unauthorized Access');
+        }
+
         $listing->delete();
         return redirect('/')->with('success', 'Listing deleted successfully!');
+    }
+
+    // Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => Listing::where('user_id', Auth::user()->id)->get()]);
     }
 }
